Split data processor nodes in several files
