+++ plugins/builtin/include/content/data_processor_nodes.hpp
@@ -1,0 +1,11 @@
+#include <hex/data_processor/node.hpp>
+
+namespace hex::plugin::builtin {
+    void registerBasicDPNs();
+    void registerVisualDPNs();
+    void registerLogicDPNs();
+    void registerControlDPNs();
+    void registerDecodeDPNs();
+    void registerMathDPNs();
+    void registerOtherDPNs();
+}
